requirements.txt
cmake
easydict~=1.9
numpy
numba
opencv-python-headless~=4.5.1.48
scipy
scikit-image
tqdm
PyYAML
open3d
cython
tensorboardX
shapely==2.0.0
einops
timm
icecream
h5py
pyquaternion
efficientnet_pytorch==0.7.0
scikit_learn
python-lzf
termcolor
torchinfo
git+https://github.com/klintan/pypcd.git